refactor(lab): move tests out of src/ into test/
Project convention: no test code in atuin-share/src. A Cargo test target is a
separate crate that sees only the public API, so the private-internal unit
tests (transport/session/compositor/keyframe/backpressure/protocol/query/
subshell) are removed rather than relocated; the surviving public-API tests
(Size, ShareOptions, the run_share tty guard) move to test/public_api.rs,
registered via an explicit [[test]] target since Cargo only auto-discovers
tests/. Also drops the now-unused visible_width helper and rstest dev-dep.

Coverage note: this removes ~34 internal unit tests, including those guarding
recent fixes (seq-gap, keyframe cadence, backpressure resync, write-gating).
They remain in git history if we want them back as integration tests behind a
pub(crate) test facade.
