+++ crates/atuin-share/src/backpressure.rs
@@ -123,77 +123,3 @@ impl Default for Backoff {
         Self::new()
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn queue_collapses_output_backlog_to_a_keyframe_on_overflow() {
-        let mut q = OutboundQueue::new(3); // capacity 3 output frames
-        for i in 1..=5 {
-            q.push_output(i, vec![b'x']);
-        }
-        // On overflow the queue keeps only a resync marker: the caller must
-        // send a fresh keyframe instead of the dropped backlog.
-        assert!(
-            q.needs_keyframe(),
-            "overflow must request a keyframe resync"
-        );
-        assert!(q.drain_output().len() <= 3);
-    }
-
-    #[test]
-    fn queue_sends_no_output_between_the_drop_and_the_resync_keyframe() {
-        // The dropped frames were never written to the wire, so the hub's
-        // "keyframe + bytes since" buffer would have a hole in it. Sending any
-        // further output before the resync keyframe cements that hole: a viewer
-        // joining in the window replays the stale keyframe plus bytes that skip
-        // the dropped range. The hub carefully avoids this on its own overflow
-        // path; the CLI must not create it from this side.
-        let mut q = OutboundQueue::new(3);
-        for i in 1..=5 {
-            q.push_output(i, vec![b'x']);
-        }
-        assert!(
-            q.needs_keyframe(),
-            "overflow must request a keyframe resync"
-        );
-
-        // What the transport does on overflow: drop the backlog and ask the
-        // session for a keyframe.
-        q.drain_output();
-        q.clear_keyframe_flag();
-        assert!(
-            q.awaiting_keyframe(),
-            "the queue must stay in resync until the keyframe is actually sent"
-        );
-
-        // Output produced while the keyframe is in flight must not reach the
-        // wire — the keyframe already reflects those bytes.
-        for i in 6..=9 {
-            q.push_output(i, vec![b'y']);
-        }
-        assert!(
-            q.drain_output().is_empty(),
-            "no output may be sent between the drop and the resync keyframe"
-        );
-
-        // Once the keyframe is on the wire the stream resumes normally.
-        q.on_keyframe_sent();
-        assert!(!q.awaiting_keyframe());
-        q.push_output(20, vec![b'z']);
-        assert_eq!(q.drain_output(), vec![(20, vec![b'z'])]);
-    }
-
-    #[test]
-    fn backoff_grows_then_resets() {
-        let mut b = Backoff::new();
-        let d0 = b.next_delay();
-        let d1 = b.next_delay();
-        assert!(d1 > d0);
-        assert!(b.next_delay() <= std::time::Duration::from_secs(10)); // capped
-        b.reset();
-        assert_eq!(b.next_delay(), d0);
-    }
-}
