+++ crates/atuin-share/src/lib.rs
@@ -179,46 +179,3 @@ pub fn run_share(_opts: ShareOptions) -> eyre::Result<()> {
         "atuin lab share currently supports unix platforms only"
     ))
 }
-
-#[cfg(all(test, unix))]
-mod run_tests {
-    use super::*;
-
-    #[test]
-    fn refuses_when_not_a_tty() {
-        // In `cargo nextest`, stdin is not a terminal. This guard is also what
-        // keeps the test suite from ever enabling raw mode on the developer's
-        // real terminal.
-        let err = run_share(ShareOptions {
-            write: false,
-            hub_url: "wss://x".into(),
-            api_token: "t".into(),
-        })
-        .unwrap_err();
-        assert!(format!("{err:#}").contains("terminal"));
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn size_is_copy_and_eq() {
-        let a = Size { cols: 80, rows: 24 };
-        let b = a;
-        assert_eq!(a, b);
-    }
-
-    #[test]
-    fn share_options_round_trip() {
-        let o = ShareOptions {
-            write: true,
-            hub_url: "wss://h".into(),
-            api_token: "tok".into(),
-        };
-        assert!(o.write);
-        assert_eq!(o.hub_url, "wss://h");
-        assert_eq!(o.api_token, "tok");
-    }
-}
