+++ crates/atuin-share/src/protocol.rs
@@ -82,60 +82,3 @@ pub fn decode(raw: &str) -> Result<Incoming, serde_json::Error> {
         },
     })
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn join_encodes_as_phoenix_v2_array() {
-        let msg = encode_join("1", "1", "share:host", serde_json::json!({ "write": true }));
-        let v: serde_json::Value = serde_json::from_str(&msg).unwrap();
-        assert_eq!(v[0], "1"); // join_ref
-        assert_eq!(v[1], "1"); // ref
-        assert_eq!(v[2], "share:host"); // topic
-        assert_eq!(v[3], "phx_join"); // event
-        assert_eq!(v[4]["write"], true); // payload
-    }
-
-    #[test]
-    fn heartbeat_uses_null_join_ref_and_phoenix_topic() {
-        let msg = encode_heartbeat("7");
-        let v: serde_json::Value = serde_json::from_str(&msg).unwrap();
-        assert!(v[0].is_null());
-        assert_eq!(v[1], "7");
-        assert_eq!(v[2], "phoenix");
-        assert_eq!(v[3], "heartbeat");
-    }
-
-    #[test]
-    fn decode_reply_ok_extracts_response() {
-        let raw = r#"["1","1","share:host","phx_reply",{"status":"ok","response":{"token":"abc","join_url":"u"}}]"#;
-        match decode(raw).unwrap() {
-            Incoming::Reply { ref_, ok, response } => {
-                assert_eq!(ref_, "1");
-                assert!(ok);
-                assert_eq!(response["token"], "abc");
-            }
-            other => panic!("expected reply, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn decode_event_extracts_event_and_payload() {
-        let raw = r#"["1","5","share:host","set_size",{"cols":80,"rows":24}]"#;
-        match decode(raw).unwrap() {
-            Incoming::Event { event, payload } => {
-                assert_eq!(event, "set_size");
-                assert_eq!(payload["rows"], 24);
-            }
-            other => panic!("expected event, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn base64_round_trips_raw_bytes() {
-        let bytes = [0u8, 27, 91, 255, 10];
-        assert_eq!(b64_decode(&b64_encode(&bytes)).unwrap(), bytes);
-    }
-}
