+++ crates/atuin-share/src/query.rs
@@ -54,38 +54,3 @@ pub fn replies(chunk: &[u8], cursor: (u16, u16)) -> Vec<u8> {
     }
     out
 }
-
-#[cfg(all(test, unix))]
-mod tests {
-    use super::replies;
-
-    #[test]
-    fn answers_cursor_position_report_1_indexed() {
-        // vt100 reports 0-indexed; the CPR reply is 1-indexed.
-        assert_eq!(replies(b"\x1b[6n", (4, 9)), b"\x1b[5;10R".to_vec());
-    }
-
-    #[test]
-    fn answers_primary_device_attributes() {
-        // Claim a plain VT102-ish terminal.
-        assert_eq!(replies(b"\x1b[c", (0, 0)), b"\x1b[?6c".to_vec());
-    }
-
-    #[test]
-    fn answers_device_attributes_with_explicit_zero_param() {
-        assert_eq!(replies(b"\x1b[0c", (0, 0)), b"\x1b[?6c".to_vec());
-    }
-
-    #[test]
-    fn ignores_ordinary_output() {
-        assert!(replies(b"hello\r\n\x1b[1;31mred\x1b[0m", (0, 0)).is_empty());
-    }
-
-    #[test]
-    fn answers_every_query_in_one_chunk_in_order() {
-        assert_eq!(
-            replies(b"a\x1b[6nb\x1b[cc", (0, 0)),
-            b"\x1b[1;1R\x1b[?6c".to_vec()
-        );
-    }
-}
