+++ crates/atuin-share/src/render.rs
@@ -12,30 +12,6 @@ use unicode_width::UnicodeWidthStr;
 
 use crate::Size;
 
-/// Display width of the printable columns, ignoring ANSI escape sequences.
-/// Test + assertion helper. Measured in *columns*, not chars: the bar text
-/// contains wide glyphs (`⚠`) and a char count would under-measure them.
-#[cfg_attr(
-    not(test),
-    allow(dead_code, reason = "assertion helper, only called from tests")
-)]
-pub(crate) fn visible_width(s: &str) -> usize {
-    let mut visible = String::with_capacity(s.len());
-    let mut in_esc = false;
-    for ch in s.chars() {
-        if in_esc {
-            if ch.is_ascii_alphabetic() {
-                in_esc = false;
-            }
-        } else if ch == '\x1b' {
-            in_esc = true;
-        } else {
-            visible.push(ch);
-        }
-    }
-    UnicodeWidthStr::width(visible.as_str())
-}
-
 /// What the bar's chunks are joined with.
 const SEPARATOR: &str = " · ";
 
@@ -160,106 +136,5 @@ impl std::fmt::Write for Utf8<'_> {
     fn write_str(&mut self, s: &str) -> std::fmt::Result {
         self.0.extend_from_slice(s.as_bytes());
         Ok(())
-    }
-}
-
-#[cfg(all(test, unix))]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn bar_is_clamped_to_width_and_mentions_state() {
-        let bar = render_bar(30, 3, true);
-        let text = String::from_utf8_lossy(&bar);
-        assert!(text.contains("SHARED SESSION"));
-        assert!(text.contains("WRITE ON"));
-        // The visible run must occupy exactly `cols` display columns: the bar is
-        // a full reverse-video row, so it is padded as well as truncated.
-        assert_eq!(super::visible_width(&text), 30);
-    }
-
-    #[test]
-    fn bar_shows_write_off_when_read_only() {
-        let text = String::from_utf8_lossy(&render_bar(80, 0, false)).into_owned();
-        assert!(text.contains("WRITE OFF"));
-        assert_eq!(super::visible_width(&text), 80);
-    }
-
-    #[test]
-    fn bar_width_is_display_columns_not_char_count() {
-        // The bar text opens with '⚠', which is a wide (2-column) glyph. A
-        // char-count clamp would overflow the row and wrap, destroying the
-        // layout, so the budget must be measured in display columns.
-        for cols in [1u16, 2, 3, 8, 40] {
-            assert_eq!(
-                super::visible_width(&String::from_utf8_lossy(&render_bar(cols, 1, true))),
-                cols as usize,
-                "bar must be exactly {cols} columns wide"
-            );
-        }
-    }
-
-    #[test]
-    fn composite_places_bar_on_row1_and_child_below() {
-        let mut p = vt100::Parser::new(4, 20, 0); // child area: 4 rows
-        p.process(b"\x1b[HABC");
-        let child = crate::Size { cols: 20, rows: 4 };
-        // Rows available below the bar; the real terminal is 5 rows (1 + 4).
-        let avail = crate::Size { cols: 20, rows: 4 };
-        let bar = render_bar(20, 1, false);
-        let frame = composite(p.screen(), child, avail, &bar);
-        let s = String::from_utf8_lossy(&frame);
-
-        // The bar is written first, anchored absolutely at row 1.
-        assert!(s.starts_with("\x1b[1;1H"));
-        // Scroll region spans the child's rows only: physical row 2 ..= 1+child.rows.
-        assert!(
-            s.contains("\x1b[2;5r"),
-            "scroll region must be 2..=5, got {s:?}"
-        );
-        // Origin mode is enabled for the child repaint, then disabled again.
-        assert!(s.contains("\x1b[?6h"));
-        assert!(s.contains("\x1b[?6l"));
-        // The cursor is finally placed at child cursor row+2 / col+1, absolute.
-        let (cr, cc) = p.screen().cursor_position();
-        assert!(s.contains(&format!("\x1b[{};{}H", cr + 2, cc + 1)));
-    }
-
-    #[test]
-    fn composite_keeps_the_last_row_when_the_child_fills_the_area() {
-        // The default state: nothing has shrunk the session, so the child is
-        // exactly as tall as the area below the bar (`avail`). The bottom margin
-        // must then be `1 + child.rows` — the last physical row. Clamping to
-        // `avail.rows` instead would subtract the bar row a second time (spec
-        // §6), leaving the region one row short: every repaint would scroll the
-        // child's top line off and the last physical row would stay blank.
-        let mut p = vt100::Parser::new(9, 20, 0);
-        p.process(b"\x1b[Hx");
-        let child = crate::Size { cols: 20, rows: 9 };
-        let avail = crate::Size { cols: 20, rows: 9 }; // real terminal: 1 bar + 9
-        let bar = render_bar(20, 0, false);
-        let s = String::from_utf8_lossy(&composite(p.screen(), child, avail, &bar)).into_owned();
-        assert!(
-            s.contains("\x1b[2;10r"),
-            "expected region 2..=10 (the child's 9 rows below the bar), got {s:?}"
-        );
-    }
-
-    #[test]
-    fn composite_clamps_child_taller_than_the_visible_area() {
-        // If the host shrinks their window, `avail` can be smaller than the last
-        // negotiated `child`. Painting `child.rows` regardless would run past
-        // the last visible row, scroll the screen, and destroy the bar.
-        let mut p = vt100::Parser::new(20, 20, 0);
-        p.process(b"\x1b[Hx");
-        let child = crate::Size { cols: 20, rows: 20 };
-        let avail = crate::Size { cols: 20, rows: 5 }; // real terminal: 1 bar + 5
-        let bar = render_bar(20, 0, false);
-        let s = String::from_utf8_lossy(&composite(p.screen(), child, avail, &bar)).into_owned();
-        // Bottom margin is clamped to the physical last row, never beyond it.
-        assert!(
-            s.contains("\x1b[2;6r"),
-            "expected clamped region 2..=6, got {s:?}"
-        );
     }
 }
