+++ crates/atuin-share/src/session.rs
@@ -663,431 +663,3 @@ impl Session {
         Ok(exit_code)
     }
 }
-
-#[cfg(all(test, unix))]
-// `std::env::set_var` is `unsafe` in edition 2024, and the crate denies
-// `unsafe_code`; the tests below set `$SHELL` before spawning any threads that
-// read the environment, so the calls are sound.
-#[allow(unsafe_code)]
-mod tests {
-    use std::io::Read;
-    use std::sync::Arc;
-    use std::sync::atomic::{AtomicBool, Ordering};
-    use std::sync::mpsc;
-    use std::time::{Duration, Instant};
-
-    use super::*;
-    use crate::Size;
-    use crate::subshell::Subshell;
-
-    /// A `Read` standing in for the host's terminal: it yields `prologue` on the
-    /// first read, then blocks until `finish` is set and yields the kill switch
-    /// (`Ctrl-\`, `0x1c`).
-    ///
-    /// Blocking on an explicit flag rather than handing the session a
-    /// pre-filled `Cursor` is what makes these tests deterministic: a `Cursor`
-    /// is drained the instant the stdin thread starts, so the kill switch would
-    /// routinely beat the child shell's very first byte to the channel and the
-    /// assertions about startup traffic would be a coin flip.
-    struct ScriptedStdin {
-        prologue: Option<Vec<u8>>,
-        finish: Arc<AtomicBool>,
-        sent_kill: bool,
-    }
-
-    impl ScriptedStdin {
-        fn new(prologue: Vec<u8>, finish: Arc<AtomicBool>) -> Self {
-            Self {
-                prologue: (!prologue.is_empty()).then_some(prologue),
-                finish,
-                sent_kill: false,
-            }
-        }
-    }
-
-    impl Read for ScriptedStdin {
-        fn read(&mut self, buf: &mut [u8]) -> std::io::Result<usize> {
-            if let Some(prologue) = self.prologue.take() {
-                let n = prologue.len().min(buf.len());
-                buf[..n].copy_from_slice(&prologue[..n]);
-                return Ok(n);
-            }
-            if self.sent_kill || buf.is_empty() {
-                return Ok(0);
-            }
-            while !self.finish.load(Ordering::SeqCst) {
-                std::thread::sleep(Duration::from_millis(5));
-            }
-            self.sent_kill = true;
-            buf[0] = 0x1c;
-            Ok(1)
-        }
-    }
-
-    #[test]
-    fn forwards_output_with_monotonic_seq_and_emits_initial_keyframe() {
-        unsafe { std::env::set_var("SHELL", "/bin/sh") };
-        let sh = Subshell::spawn(Size { cols: 40, rows: 9 }, &[]).unwrap();
-        let (out_tx, out_rx) = mpsc::channel::<Outbound>();
-        let (in_tx, in_rx) = mpsc::channel::<Inbound>();
-
-        // Kill switch once the startup traffic has been observed.
-        let finish = Arc::new(AtomicBool::new(false));
-        let stdin = ScriptedStdin::new(b"printf HELLO\n".to_vec(), Arc::clone(&finish));
-        let stdout = Vec::new();
-
-        let handle = std::thread::spawn(move || {
-            Session::run(
-                sh,
-                Size { cols: 40, rows: 9 },
-                false,
-                out_tx,
-                in_rx,
-                Box::new(stdin),
-                Box::new(stdout),
-            )
-        });
-        let _ = in_tx.send(Inbound::Connected {
-            token: "t".into(),
-            join_url: "u".into(),
-            fresh_session: false,
-        });
-
-        let mut seqs = Vec::new();
-        let mut saw_keyframe = false;
-        let mut saw_host_size = false;
-        let mut saw_end = false;
-        let deadline = Instant::now() + Duration::from_secs(30);
-        while Instant::now() < deadline {
-            match out_rx.recv_timeout(Duration::from_secs(5)) {
-                Ok(Outbound::Keyframe { seq, .. }) => {
-                    saw_keyframe = true;
-                    seqs.push(seq);
-                }
-                Ok(Outbound::Output { seq, .. }) => seqs.push(seq),
-                Ok(Outbound::HostSize { .. }) => saw_host_size = true,
-                Ok(Outbound::End) => {
-                    saw_end = true;
-                    break;
-                }
-                Err(_) => break,
-            }
-            if seqs.len() >= 2 && saw_keyframe && saw_host_size {
-                finish.store(true, Ordering::SeqCst);
-            }
-        }
-        finish.store(true, Ordering::SeqCst);
-        let _ = handle.join();
-
-        assert!(saw_keyframe, "expected an initial keyframe");
-        assert!(saw_host_size, "expected host_size to be sent at startup");
-        assert!(
-            saw_end,
-            "expected an explicit end event from the kill switch"
-        );
-        // strictly increasing
-        assert!(
-            seqs.windows(2).all(|w| w[1] > w[0]),
-            "seqs not monotonic: {seqs:?}"
-        );
-    }
-
-    #[test]
-    fn read_only_session_discards_viewer_input() {
-        // Defence in depth (spec §8): even if the hub forwards `input` for a
-        // read-only session, the CLI must never write it to the child.
-        unsafe { std::env::set_var("SHELL", "/bin/sh") };
-        let sh = Subshell::spawn(Size { cols: 40, rows: 9 }, &[]).unwrap();
-        let (out_tx, out_rx) = mpsc::channel::<Outbound>();
-        let (in_tx, in_rx) = mpsc::channel::<Inbound>();
-
-        let finish = Arc::new(AtomicBool::new(false));
-        let stdin = ScriptedStdin::new(Vec::new(), Arc::clone(&finish));
-
-        let handle = std::thread::spawn(move || {
-            Session::run(
-                sh,
-                Size { cols: 40, rows: 9 },
-                false, // read-only
-                out_tx,
-                in_rx,
-                Box::new(stdin),
-                Box::new(Vec::new()),
-            )
-        });
-
-        // A command that would print a distinctive marker if it were executed.
-        let _ = in_tx.send(Inbound::Input(b"printf VIEWERWROTE\n".to_vec()));
-
-        let deadline = Instant::now() + Duration::from_secs(3);
-        let mut seen = String::new();
-        while Instant::now() < deadline {
-            if let Ok(Outbound::Output { data, .. }) =
-                out_rx.recv_timeout(Duration::from_millis(200))
-            {
-                seen.push_str(&String::from_utf8_lossy(&data));
-            }
-        }
-        // The viewer cannot close this shell — its keystrokes are dropped — so
-        // the host's own kill switch is what ends the session.
-        finish.store(true, Ordering::SeqCst);
-        let _ = handle.join();
-
-        assert!(
-            !seen.contains("VIEWERWROTE"),
-            "read-only session executed viewer input: {seen:?}"
-        );
-    }
-
-    #[test]
-    fn set_size_resizes_the_child_and_forces_a_keyframe() {
-        unsafe { std::env::set_var("SHELL", "/bin/sh") };
-        let sh = Subshell::spawn(Size { cols: 40, rows: 9 }, &[]).unwrap();
-        let (out_tx, out_rx) = mpsc::channel::<Outbound>();
-        let (in_tx, in_rx) = mpsc::channel::<Inbound>();
-
-        let finish = Arc::new(AtomicBool::new(false));
-        let stdin = ScriptedStdin::new(Vec::new(), Arc::clone(&finish));
-
-        let handle = std::thread::spawn(move || {
-            Session::run(
-                sh,
-                Size { cols: 40, rows: 9 },
-                true,
-                out_tx,
-                in_rx,
-                Box::new(stdin),
-                Box::new(Vec::new()),
-            )
-        });
-
-        // Drain the startup keyframe first. `want_keyframe` is a single flag, so
-        // a resize requested before the initial keyframe has been serviced would
-        // simply be absorbed by it and prove nothing.
-        let _ = in_tx.send(Inbound::Input(b"printf READY\n".to_vec()));
-        let mut startup_keyframes = 0usize;
-        let startup_deadline = Instant::now() + Duration::from_secs(15);
-        while Instant::now() < startup_deadline && startup_keyframes == 0 {
-            if let Ok(Outbound::Keyframe { .. }) = out_rx.recv_timeout(Duration::from_millis(200)) {
-                startup_keyframes += 1;
-            }
-        }
-        assert_eq!(startup_keyframes, 1, "expected an initial keyframe");
-
-        let _ = in_tx.send(Inbound::SetSize { cols: 30, rows: 6 });
-        // The child observes the new size; `stty size` prints "rows cols".
-        let _ = in_tx.send(Inbound::Input(b"stty size\n".to_vec()));
-
-        let deadline = Instant::now() + Duration::from_secs(15);
-        let mut seen = String::new();
-        let mut keyframes = 0usize;
-        while Instant::now() < deadline && !seen.contains("6 30") {
-            match out_rx.recv_timeout(Duration::from_millis(200)) {
-                Ok(Outbound::Output { data, .. }) => {
-                    seen.push_str(&String::from_utf8_lossy(&data));
-                }
-                Ok(Outbound::Keyframe { .. }) => keyframes += 1,
-                Ok(_) | Err(_) => {}
-            }
-        }
-        let _ = in_tx.send(Inbound::Input(vec![0x04]));
-        finish.store(true, Ordering::SeqCst);
-        let _ = handle.join();
-
-        assert!(
-            seen.contains("6 30"),
-            "child did not see the new size: {seen:?}"
-        );
-        assert!(
-            keyframes >= 1,
-            "expected a keyframe after resize, saw {keyframes}"
-        );
-    }
-
-    #[test]
-    fn resize_notice_is_silent_on_a_no_op_and_honest_about_the_cause() {
-        // The hub's first `set_size` normally just echoes the host's own size,
-        // which used to print
-        //   "resized to 100×29 — a viewer's screen is smaller"
-        // at startup with nobody watching. Both halves were false: nothing
-        // resized, and no viewer caused it.
-        let current = Size {
-            cols: 100,
-            rows: 29,
-        };
-        assert_eq!(
-            resize_notice(current, current, 0),
-            None,
-            "a no-op set_size must not announce a resize"
-        );
-        assert_eq!(
-            resize_notice(current, current, 3),
-            None,
-            "a no-op set_size must stay silent even with viewers attached"
-        );
-
-        // A real change with viewers attached: they are the reason.
-        let smaller = Size { cols: 80, rows: 24 };
-        let with_viewers = resize_notice(current, smaller, 2).expect("a real resize must announce");
-        assert!(with_viewers.contains("80×24"), "got {with_viewers:?}");
-        assert!(
-            with_viewers.contains("viewer"),
-            "should attribute to a viewer: {with_viewers:?}"
-        );
-
-        // A real change with nobody watching cannot be blamed on a viewer.
-        let no_viewers = resize_notice(current, smaller, 0).expect("a real resize must announce");
-        assert!(no_viewers.contains("80×24"), "got {no_viewers:?}");
-        assert!(
-            !no_viewers.contains("viewer"),
-            "must not invent a viewer when none are connected: {no_viewers:?}"
-        );
-    }
-
-    #[test]
-    fn emits_keyframe_on_set_size_even_when_the_child_is_idle() {
-        // Spec §5.3 requires a keyframe after EVERY `set_size`. If the request
-        // is only serviced by the PTY read loop, an idle child never produces
-        // one: viewers are told the new `size`, resize their grid, and then wait
-        // forever for the matching repaint.
-        //
-        // The child is `cat`, not a shell, and that is load-bearing: resizing the
-        // PTY raises SIGWINCH in the child, and an interactive shell answers it
-        // by redrawing its prompt — output that would service the keyframe
-        // request by accident and hide the bug. `cat` ignores SIGWINCH and never
-        // writes unless written to, so the resize is the only possible cause.
-        unsafe { std::env::set_var("SHELL", "/bin/cat") };
-        let sh = Subshell::spawn(Size { cols: 40, rows: 9 }, &[]).unwrap();
-        let (out_tx, out_rx) = mpsc::channel::<Outbound>();
-        let (in_tx, in_rx) = mpsc::channel::<Inbound>();
-
-        let finish = Arc::new(AtomicBool::new(false));
-        let stdin = ScriptedStdin::new(Vec::new(), Arc::clone(&finish));
-
-        let handle = std::thread::spawn(move || {
-            Session::run(
-                sh,
-                Size { cols: 40, rows: 9 },
-                true,
-                out_tx,
-                in_rx,
-                Box::new(stdin),
-                Box::new(Vec::new()),
-            )
-        });
-
-        // Drain startup traffic (host_size, any prompt, the initial keyframe)
-        // and wait for the channel to fall quiet, so what follows can only have
-        // been caused by the resize.
-        let settle = Instant::now() + Duration::from_secs(3);
-        while Instant::now() < settle {
-            if out_rx.recv_timeout(Duration::from_millis(400)).is_err() {
-                break; // quiet — the child is idle
-            }
-        }
-
-        let _ = in_tx.send(Inbound::SetSize { cols: 30, rows: 6 });
-
-        // Well under the 5 s periodic cadence, so only an immediate emit passes.
-        let mut keyframe = None;
-        let deadline = Instant::now() + Duration::from_millis(1500);
-        while Instant::now() < deadline {
-            match out_rx.recv_timeout(Duration::from_millis(100)) {
-                Ok(Outbound::Keyframe { data, .. }) => {
-                    keyframe = Some(data);
-                    break;
-                }
-                Ok(_) | Err(_) => {}
-            }
-        }
-        finish.store(true, Ordering::SeqCst);
-        let _ = handle.join();
-
-        let data = keyframe.expect("no keyframe followed set_size on an idle child");
-        // The repaint must apply cleanly to the newly negotiated geometry.
-        let mut term = vt100::Parser::new(6, 30, 0);
-        term.process(&data);
-        assert_eq!(term.screen().size(), (6, 30));
-    }
-
-    /// A `Write` the test can read back while the renderer thread owns it.
-    #[derive(Clone, Default)]
-    struct SharedSink(Arc<Mutex<Vec<u8>>>);
-
-    impl SharedSink {
-        fn snapshot(&self) -> Vec<u8> {
-            self.0.lock().expect("sink lock").clone()
-        }
-    }
-
-    impl Write for SharedSink {
-        fn write(&mut self, buf: &[u8]) -> std::io::Result<usize> {
-            self.0.lock().expect("sink lock").extend_from_slice(buf);
-            Ok(buf.len())
-        }
-        fn flush(&mut self) -> std::io::Result<()> {
-            Ok(())
-        }
-    }
-
-    #[test]
-    fn renderer_paints_every_child_row_below_the_bar() {
-        // Regression: the bar row is subtracted exactly once, by `run_share`,
-        // so `Session::run` is handed `rows = real_rows - 1` and `Shared`
-        // stores that already-subtracted value. If the compositor then clamps
-        // the scroll region to it a second time, the region is one row short of
-        // the child, every frame scrolls the child's top line away, and the last
-        // physical row is never painted (spec §6, §7).
-        //
-        // This is the *default* state — no viewer has shrunk the session — so it
-        // is what the host sees on a plain `atuin lab share`.
-        let real_rows: u16 = 10;
-        let physical = Size {
-            cols: 20,
-            rows: real_rows - 1, // exactly what `run_share` passes in
-        };
-        let shared = Arc::new(Shared::new(physical));
-        {
-            let mut parser = shared.parser.lock().expect("parser lock");
-            for r in 0..physical.rows {
-                parser.process(format!("\x1b[{};1HL{r}", r + 1).as_bytes());
-            }
-        }
-
-        let sink = SharedSink::default();
-        let renderer = spawn_renderer(Arc::clone(&shared), Box::new(sink.clone()), false);
-
-        let deadline = Instant::now() + Duration::from_secs(5);
-        let mut frame = Vec::new();
-        while Instant::now() < deadline {
-            frame = sink.snapshot();
-            if !frame.is_empty() {
-                break;
-            }
-            std::thread::sleep(Duration::from_millis(5));
-        }
-        shared.shutdown.store(true, Ordering::SeqCst);
-        let _ = renderer.join();
-        assert!(!frame.is_empty(), "renderer never painted a frame");
-
-        // Replay the exact bytes through a terminal the size of the host's REAL
-        // screen — bar row included.
-        let mut term = vt100::Parser::new(real_rows, physical.cols, 0);
-        term.process(&frame);
-        let rows: Vec<String> = term.screen().rows(0, physical.cols).collect();
-
-        assert!(
-            rows[0].contains("SHARED SESSION"),
-            "row 0 must hold the bar, got {:?}",
-            rows[0]
-        );
-        for r in 0..physical.rows {
-            assert_eq!(
-                rows[usize::from(r) + 1].trim_end(),
-                format!("L{r}"),
-                "child row {r} must land on physical row {}; full frame: {rows:?}",
-                r + 1
-            );
-        }
-    }
-}
