+++ crates/atuin-share/src/subshell.rs
@@ -166,98 +166,3 @@ impl Subshell {
         Ok(i32::try_from(status.exit_code()).unwrap_or(1))
     }
 }
-
-#[cfg(all(test, unix))]
-// `std::env::set_var` is `unsafe` in edition 2024, and the crate denies
-// `unsafe_code`; the test below is single-threaded so the call is sound.
-#[allow(unsafe_code)]
-mod tests {
-    use std::io::Read;
-
-    use super::*;
-    use crate::Size;
-
-    #[test]
-    fn spawns_shell_and_reads_output() {
-        // Force a known program via $SHELL so the test is deterministic.
-        // SAFETY: single-threaded test; std::env::set_var is fine here.
-        unsafe { std::env::set_var("SHELL", "/bin/sh") };
-        let sh = Subshell::spawn(Size { cols: 40, rows: 10 }, &[]).unwrap();
-        {
-            use std::io::Write;
-            sh.writer().write_all(b"printf ATUINOK\n").unwrap();
-        }
-        let mut reader = sh.reader();
-        let mut buf = [0u8; 4096];
-        let mut seen = String::new();
-        for _ in 0..50 {
-            if let Ok(n) = reader.read(&mut buf) {
-                if n == 0 {
-                    break;
-                }
-                seen.push_str(&String::from_utf8_lossy(&buf[..n]));
-                if seen.contains("ATUINOK") {
-                    break;
-                }
-            }
-        }
-        assert!(
-            seen.contains("ATUINOK"),
-            "did not observe command output: {seen:?}"
-        );
-    }
-
-    #[test]
-    fn spawn_sets_env_and_preserves_inherited_env() {
-        // `new_default_prog` spawns the login shell from the passwd database
-        // (not $SHELL), so we can't force a specific shell here. Instead we run
-        // the probe then `exit` and read to EOF, which is deterministic for any
-        // POSIX shell regardless of prompt/line-editing noise. The expanded
-        // `U=...` output is emitted by `printf` on its own line, so it is not
-        // corrupted by input-echo line wrapping.
-        //
-        // SAFETY: single-threaded test; std::env::set_var is fine here.
-        // A var set in the PARENT the child must still inherit.
-        unsafe { std::env::set_var("ATUIN_SHARE_TEST_INHERIT", "yes") };
-
-        let sh = Subshell::spawn(
-            Size { cols: 80, rows: 24 },
-            &[("ATUIN_SHARE_URL", "http://h/lab/share/tok")],
-        )
-        .unwrap();
-        {
-            use std::io::Write;
-            sh.writer()
-                .write_all(
-                    b"printf '\\nU=%s I=%s\\n' \"$ATUIN_SHARE_URL\" \"$ATUIN_SHARE_TEST_INHERIT\"; exit\n",
-                )
-                .unwrap();
-        }
-        let mut reader = sh.reader();
-        let mut buf = [0u8; 4096];
-        let mut seen = String::new();
-        // Bounded so a misbehaving shell fails the test instead of hanging;
-        // normally the loop ends at EOF/EIO once the child has exited.
-        for _ in 0..1000 {
-            match reader.read(&mut buf) {
-                Ok(0) => break,
-                Ok(n) => {
-                    seen.push_str(&String::from_utf8_lossy(&buf[..n]));
-                    if seen.contains("U=http://h/lab/share/tok") && seen.contains("I=yes") {
-                        break;
-                    }
-                }
-                // EIO is the normal way a PTY master reports the child exiting.
-                Err(_) => break,
-            }
-        }
-        assert!(
-            seen.contains("U=http://h/lab/share/tok"),
-            "ATUIN_SHARE_URL was not set in the child: {seen:?}"
-        );
-        assert!(
-            seen.contains("I=yes"),
-            "child did not inherit the parent environment: {seen:?}"
-        );
-    }
-}
