+++ crates/atuin-share/src/transport.rs
@@ -432,174 +432,3 @@ pub fn spawn_transport(
         runtime.block_on(client.run(&mut bridge_rx, &in_tx));
     });
 }
-
-#[cfg(test)]
-mod tests {
-    use std::sync::mpsc;
-
-    use super::*;
-
-    #[test]
-    fn builds_phoenix_ws_url() {
-        let u = ws_url("wss://hub.atuin.sh", "APITOKEN");
-        assert_eq!(
-            u,
-            "wss://hub.atuin.sh/sockets/share/websocket?vsn=2.0.0&token=APITOKEN"
-        );
-    }
-
-    #[test]
-    fn join_payload_is_create_then_resume() {
-        let create = join_payload(true, None);
-        assert_eq!(create["write"], true);
-        assert!(create.get("resume_token").is_none());
-
-        // On reconnect the payload carries the SECRET host_resume_token — never
-        // the public share token.
-        let resume = join_payload(true, Some("secret-host-token"));
-        assert_eq!(resume["resume_token"], "secret-host-token");
-        assert_eq!(resume["write"], true);
-    }
-
-    /// A `Client` with no socket, for exercising the join-reply bookkeeping.
-    fn test_client() -> Client {
-        Client {
-            hub_url: "ws://localhost:4000".into(),
-            api_token: "atapi_test".into(),
-            write: false,
-            host_resume_token: None,
-            last_public_token: None,
-            first_url_tx: None,
-            backoff: Backoff::new(),
-        }
-    }
-
-    #[test]
-    fn first_join_reports_the_url_exactly_once() {
-        use std::time::Duration;
-
-        let (url_tx, url_rx) = mpsc::channel::<String>();
-        let mut client = Client {
-            first_url_tx: Some(url_tx),
-            ..test_client()
-        };
-        let (in_tx, _in_rx) = mpsc::channel::<Inbound>();
-
-        let reply = json!({
-            "token": "pub1",
-            "join_url": "http://h/lab/share/pub1",
-            "host_resume_token": "secret1"
-        });
-        client.on_joined(&reply, &in_tx);
-
-        // A reconnect (second join) must NOT push another URL: the shell's
-        // environment was fixed at spawn and can't be changed now.
-        let reply2 = json!({
-            "token": "pub2",
-            "join_url": "http://h/lab/share/pub2",
-            "host_resume_token": "secret2"
-        });
-        client.on_joined(&reply2, &in_tx);
-
-        assert_eq!(
-            url_rx.recv_timeout(Duration::from_secs(1)).unwrap(),
-            "http://h/lab/share/pub1"
-        );
-        assert!(
-            url_rx.try_recv().is_err(),
-            "the join URL must be reported exactly once"
-        );
-    }
-
-    #[test]
-    fn first_connect_is_not_reported_as_a_replaced_session() {
-        // The hub always mints a session on a first (non-resume) join, so a
-        // naive "token changed?" check is true here and the very first thing a
-        // new user sees is "Reconnected as a NEW session — the previous link is
-        // dead". There was no previous link; the warning is both wrong and
-        // alarming. It must only fire when a resume was actually rejected.
-        let (tx, rx) = mpsc::channel::<Inbound>();
-        let mut client = test_client();
-
-        client.on_joined(
-            &json!({ "token": "pub1", "join_url": "u/pub1", "host_resume_token": "secret1" }),
-            &tx,
-        );
-
-        match rx.try_recv() {
-            Ok(Inbound::Connected { fresh_session, .. }) => assert!(
-                !fresh_session,
-                "a first connect must not be announced as a replaced session"
-            ),
-            other => panic!("expected Connected, got {:?}", other.is_ok()),
-        }
-    }
-
-    #[test]
-    fn reconnect_with_a_changed_token_is_reported_as_a_replaced_session() {
-        // A rejected/expired resume makes the hub silently mint a new session
-        // with a new public token. The old link is dead, so this one *must*
-        // warn — otherwise the host keeps advertising a URL nobody can join.
-        let (tx, rx) = mpsc::channel::<Inbound>();
-        let mut client = test_client();
-
-        client.on_joined(
-            &json!({ "token": "pub1", "join_url": "u/pub1", "host_resume_token": "secret1" }),
-            &tx,
-        );
-        let _ = rx.try_recv();
-
-        // Same session resumed: same public token, no warning.
-        client.on_joined(
-            &json!({ "token": "pub1", "join_url": "u/pub1", "host_resume_token": "secret1" }),
-            &tx,
-        );
-        match rx.try_recv() {
-            Ok(Inbound::Connected { fresh_session, .. }) => {
-                assert!(!fresh_session, "resuming the same session must not warn");
-            }
-            other => panic!("expected Connected, got {:?}", other.is_ok()),
-        }
-
-        // Resume rejected: the hub hands back a different public token.
-        client.on_joined(
-            &json!({ "token": "pub2", "join_url": "u/pub2", "host_resume_token": "secret2" }),
-            &tx,
-        );
-        match rx.try_recv() {
-            Ok(Inbound::Connected {
-                fresh_session,
-                join_url,
-                ..
-            }) => {
-                assert!(fresh_session, "a replaced session must warn");
-                assert_eq!(join_url, "u/pub2", "the new link must be the one printed");
-            }
-            other => panic!("expected Connected, got {:?}", other.is_ok()),
-        }
-    }
-
-    #[test]
-    fn maps_hub_events_onto_inbound() {
-        match to_inbound("input", &json!({ "data": b64_encode(b"ls\n") })) {
-            Some(Inbound::Input(bytes)) => assert_eq!(bytes, b"ls\n"),
-            _ => panic!("input must decode to raw bytes"),
-        }
-        match to_inbound("set_size", &json!({ "cols": 80, "rows": 24 })) {
-            Some(Inbound::SetSize { cols, rows }) => {
-                assert_eq!((cols, rows), (80, 24));
-            }
-            _ => panic!("set_size must map to SetSize"),
-        }
-        match to_inbound("participants", &json!({ "count": 3 })) {
-            Some(Inbound::Participants(n)) => assert_eq!(n, 3),
-            _ => panic!("participants must map to Participants"),
-        }
-        assert!(matches!(
-            to_inbound("request_keyframe", &json!({})),
-            Some(Inbound::RequestKeyframe)
-        ));
-        // Unknown events are ignored, not fatal (forward compatibility).
-        assert!(to_inbound("something_new", &json!({})).is_none());
-    }
-}
